pyproject.toml
[project]
name = "ghovax-blog"
version = "1.0.0"
description = "Giovanni's personal blog"
requires-python = ">=3.9,<3.13"
dependencies = [
    "beautifulsoup4>=4.12.2",
    "Flask>=2.3.3",
    "lxml>=5.0.0",
    "requests>=2.31.0",
    "werkzeug>=2.3.7",
    "feedwerk>=1.1.0",
    "Pillow>=10.2.0",
    "python-dotenv>=1.0.0",
    "python-slugify>=8.0.1",
    "sqlalchemy>=2.0.21",
    "humanize>=4.8.0",
    "Jinja2>=3.0.0",
]
